[CS] Do away with autoloading
Had to move Service to Sequella::Service to cleanly resolve circular dependencies
