+++ sequella.gemspec
@@ -18,7 +18,6 @@ Gem::Specification.new do |s|
   s.require_paths = ["lib"]
 
   s.add_runtime_dependency %q<adhearsion>, ["~> 2.1"]
-  s.add_runtime_dependency %q<activesupport>, [">= 3.0.10"]
   s.add_runtime_dependency %q<sequel>, [">= 3.40.0"]
 
   s.add_development_dependency %q<bundler>, ["~> 1.0"]
